Use psr http request in ServiceProxy
