+++ src/Controller/RequestHandler.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Controller;
 
 use App\Exception\ServiceException;
+use App\ServiceProxy\RequestFactory;
 use App\ServiceProxy\Service;
 use App\ServiceProxy\ServiceProxy;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,6 +15,7 @@ use Symfony\Component\Routing\Attribute\Route;
 readonly class RequestHandler
 {
     public function __construct(
+        private RequestFactory $requestFactory,
         private ServiceProxy $serviceProxy,
     ) {
     }
@@ -24,6 +26,23 @@ readonly class RequestHandler
     #[Route(path: '/{serviceName<[a-z]+>}{action<.+>}')]
     public function handle(Service $service, Request $request): Response
     {
-        return $this->serviceProxy->proxy($service, $request);
+        return $this->serviceProxy->proxy(
+            $service,
+            $this->requestFactory->create($service, $request),
+            $this->getAcceptableContentTypes($request)
+        );
+    }
+
+    /**
+     * @return string[]
+     */
+    private function getAcceptableContentTypes(Request $request): array
+    {
+        $acceptableContentTypes = $request->getAcceptableContentTypes();
+        if ([] === $acceptableContentTypes) {
+            $acceptableContentTypes = ['application/json'];
+        }
+
+        return $acceptableContentTypes;
     }
 }
